Rename .react.js -> .jsx, import css files
