+++ app/components/Settings/scheduler.jsx
@@ -7,8 +7,8 @@ import matchSorter from 'match-sorter';
 import Highlight from 'react-highlight';
 
 import { Link } from '../Link.react.js';
-import { Row, Column } from '../Layout.react.js';
-import Modal from '../Modal.react.js';
+import { Row, Column } from '../layout.jsx';
+import Modal from '../modal.jsx';
 import { plotlyUrl } from '../../utils/utils.js';
 
 const NO_OP = () => {};
