+++ static/index.html
@@ -12,7 +12,6 @@
     <link rel="stylesheet" type="text/css" href="static/css/Link.css"></link>
     <link rel="stylesheet" type="text/css" href="static/css/Logger.css"></link>
     <link rel="stylesheet" type="text/css" href="static/css/Preview.css"></link>
-    <link rel="stylesheet" type="text/css" href="static/css/Scheduler.css"></link>
     <link rel="stylesheet" type="text/css" href="static/css/Settings.css"></link>
     <link rel="stylesheet" type="text/css" href="static/css/TableDropdown.css"></link>
     <link rel="stylesheet" type="text/css" href="static/css/Tabs.css"></link>
